Move settings to settings folder
