+++ src/easydynamics/convolution/__init__.py
@@ -2,6 +2,5 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 from easydynamics.convolution.convolution import Convolution
-from easydynamics.convolution.convolution_settings import ConvolutionSettings
 
-__all__ = ['Convolution', 'ConvolutionSettings']
+__all__ = ['Convolution']
