+++ src/easydynamics/convolution/numerical_convolution.py
@@ -6,11 +6,11 @@ import scipp as sc
 from easyscience.variable import Parameter
 from scipy.signal import fftconvolve
 
-from easydynamics.convolution.convolution_settings import ConvolutionSettings
 from easydynamics.convolution.numerical_convolution_base import NumericalConvolutionBase
 from easydynamics.sample_model.component_collection import ComponentCollection
 from easydynamics.sample_model.components.model_component import ModelComponent
-from easydynamics.sample_model.detailed_balance_settings import DetailedBalanceSettings
+from easydynamics.settings.convolution_settings import ConvolutionSettings
+from easydynamics.settings.detailed_balance_settings import DetailedBalanceSettings
 from easydynamics.utils.detailed_balance import detailed_balance_factor
 from easydynamics.utils.utils import Numeric
