+++ src/easydynamics/convolution/numerical_convolution_base.py
@@ -8,11 +8,11 @@ import scipp as sc
 from easyscience.variable import Parameter
 
 from easydynamics.convolution.convolution_base import ConvolutionBase
-from easydynamics.convolution.convolution_settings import ConvolutionSettings
 from easydynamics.convolution.energy_grid import EnergyGrid
 from easydynamics.sample_model.component_collection import ComponentCollection
 from easydynamics.sample_model.components.model_component import ModelComponent
-from easydynamics.sample_model.detailed_balance_settings import DetailedBalanceSettings
+from easydynamics.settings.convolution_settings import ConvolutionSettings
+from easydynamics.settings.detailed_balance_settings import DetailedBalanceSettings
 from easydynamics.utils.utils import Numeric
 
 # The thresholds are illustrated in
