+++ src/easydynamics/settings/__init__.py
@@ -1,13 +1,10 @@
 # SPDX-FileCopyrightText: 2025 EasyScience contributors <https://github.com/easyscience>
 # SPDX-License-Identifier: BSD-3-Clause
-"""EasyDynamics library."""
 
-from easydynamics.analysis import Analysis
-from easydynamics.experiment import Experiment
 from easydynamics.settings.convolution_settings import ConvolutionSettings
+from easydynamics.settings.detailed_balance_settings import DetailedBalanceSettings
 
 __all__ = [
-    'Analysis',
     'ConvolutionSettings',
-    'Experiment',
+    'DetailedBalanceSettings',
 ]
