+++ src/easydynamics/settings/convolution_settings.py
